+++ src/collision/PolyhedronMesh.cpp
@@ -29,6 +29,7 @@
 #include <reactphysics3d/collision/PolygonVertexArray.h>
 #include <reactphysics3d/utils/DefaultLogger.h>
 #include <reactphysics3d/engine/PhysicsCommon.h>
+#include <reactphysics3d/utils/Error.h>
 #include <cstdlib>
 
 using namespace reactphysics3d;
@@ -39,60 +40,84 @@ using namespace reactphysics3d;
  * Create a polyhedron mesh given an array of polygons.
  * @param polygonVertexArray Pointer to the array of polygons and their vertices
  */
-PolyhedronMesh::PolyhedronMesh(PolygonVertexArray* polygonVertexArray, MemoryAllocator& allocator, bool releasePolygonVertexArray)
-               : mMemoryAllocator(allocator), mHalfEdgeStructure(allocator, polygonVertexArray->getNbFaces(), polygonVertexArray->getNbVertices(),
-                                    (polygonVertexArray->getNbFaces() + polygonVertexArray->getNbVertices() - 2) * 2), mFacesNormals(nullptr),
-                 mReleasePolygonVertexArray(releasePolygonVertexArray) {
+PolyhedronMesh::PolyhedronMesh(MemoryAllocator& allocator, uint32 nbVertices, uint32 nbFaces)
+               : mMemoryAllocator(allocator), mHalfEdgeStructure(allocator, nbFaces, nbVertices, (nbFaces + nbVertices - 2) * 2),
+                 mVertices(allocator, nbVertices), mFacesNormals(allocator, nbFaces) {
 
-   mPolygonVertexArray = polygonVertexArray;
-}
-
-// Destructor
-PolyhedronMesh::~PolyhedronMesh() {
-
-    if (mFacesNormals != nullptr) {
-
-        for (uint32 f=0; f < mHalfEdgeStructure.getNbFaces(); f++) {
-            mFacesNormals[f].~Vector3();
-        }
-
-        mMemoryAllocator.release(mFacesNormals, mHalfEdgeStructure.getNbFaces() * sizeof(Vector3));
-    }
 }
 
 /// Static factory method to create a polyhedron mesh. This methods returns null_ptr if the mesh is not valid
-PolyhedronMesh* PolyhedronMesh::create(PolygonVertexArray* polygonVertexArray, MemoryAllocator& polyhedronMeshAllocator, MemoryAllocator& dataAllocator,
-                                       bool releasePolygonVertexArray) {
+PolyhedronMesh* PolyhedronMesh::create(PolygonVertexArray* polygonVertexArray, MemoryAllocator& allocator,
+                                       std::vector<Error>& errors) {
 
-    PolyhedronMesh* mesh = new (polyhedronMeshAllocator.allocate(sizeof(PolyhedronMesh))) PolyhedronMesh(polygonVertexArray, dataAllocator,
-                                                                                                         releasePolygonVertexArray);
-
+    PolyhedronMesh* mesh = new (allocator.allocate(sizeof(PolyhedronMesh))) PolyhedronMesh(allocator,
+                                                                                           polygonVertexArray->getNbVertices(),
+                                                                                           polygonVertexArray->getNbFaces());
     // Create the half-edge structure of the mesh
-    bool isValid = mesh->createHalfEdgeStructure();
+    bool isValid = mesh->init(polygonVertexArray, errors);
 
-    if (isValid) {
-
-        // Compute the faces normals
-        mesh->computeFacesNormals();
-
-        // Compute the centroid
-        mesh->computeCentroid();
-    }
-    else {
+    if (!isValid) {
         mesh->~PolyhedronMesh();
-        polyhedronMeshAllocator.release(mesh, sizeof(PolyhedronMesh));
+        allocator.release(mesh, sizeof(PolyhedronMesh));
         mesh = nullptr;
     }
 
     return mesh;
 }
 
+// Initialize a mesh and returns errors if any
+bool PolyhedronMesh::init(PolygonVertexArray* polygonVertexArray, std::vector<Error>& errors) {
+
+    bool isValid = true;
+
+    // Copy the vertices from the PolygonVertexArray into the mesh
+    isValid &= copyVertices(polygonVertexArray, errors);
+
+    // Create the half-edge structure of the mesh
+    isValid &= createHalfEdgeStructure(polygonVertexArray, errors);
+
+    // Compute the faces normals
+    isValid &= computeFacesNormals(errors);
+
+   return isValid;
+}
+
+// Copy the vertices into the mesh
+bool PolyhedronMesh::copyVertices(PolygonVertexArray* polygonVertexArray, std::vector<Error>& errors) {
+
+    bool isValid = true;
+
+    mCentroid.setToZero();
+
+    for (uint32 i=0 ; i < polygonVertexArray->getNbVertices(); i++) {
+
+        const Vector3 vertex = polygonVertexArray->getVertex(i);
+        mVertices.add(vertex);
+        mCentroid += vertex;
+    }
+
+    if (getNbVertices() > 0) {
+
+        mCentroid /= static_cast<decimal>(getNbVertices());
+    }
+    else {
+
+        errors.push_back(Error("The mesh does not have any vertices"));
+
+        isValid = false;
+    }
+
+    return isValid;
+}
+
 // Create the half-edge structure of the mesh
 /// This method returns true if the mesh is valid or false otherwise
-bool PolyhedronMesh::createHalfEdgeStructure() {
+bool PolyhedronMesh::createHalfEdgeStructure(PolygonVertexArray* polygonVertexArray, std::vector<Error>& errors) {
+
+    bool isValid = true;
 
     // For each vertex of the mesh
-    const uint32 nbVertices = mPolygonVertexArray->getNbVertices();
+    const uint32 nbVertices = polygonVertexArray->getNbVertices();
     for (uint32 v=0; v < nbVertices; v++) {
         mHalfEdgeStructure.addVertex(v);
     }
@@ -100,22 +125,32 @@ bool PolyhedronMesh::createHalfEdgeStructure() {
     uint32 nbEdges = 0;
 
     // For each polygon face of the mesh
-    const uint32 nbFaces = mPolygonVertexArray->getNbFaces();
+    const uint32 nbFaces = polygonVertexArray->getNbFaces();
     for (uint32 f=0; f < nbFaces; f++) {
 
         // Get the polygon face
-        PolygonVertexArray::PolygonFace* face = mPolygonVertexArray->getPolygonFace(f);
+        PolygonVertexArray::PolygonFace* face = polygonVertexArray->getPolygonFace(f);
 
         Array<uint32> faceVertices(mMemoryAllocator, face->nbVertices);
 
         // For each vertex of the face
         for (uint32 v=0; v < face->nbVertices; v++) {
-            faceVertices.add(mPolygonVertexArray->getVertexIndexInFace(f, v));
+            faceVertices.add(polygonVertexArray->getVertexIndexInFace(f, v));
         }
 
         nbEdges += face->nbVertices;
 
-        assert(faceVertices.size() >= 3);
+        // If the face has less than three vertices
+        if (faceVertices.size() < 3) {
+
+            // Create a new error
+            errors.push_back(Error(std::string("The face with index " + std::to_string(f) + " has less than three vertices")));
+
+            isValid = false;
+
+            RP3D_LOG("PhysicsCommon", Logger::Level::Error, Logger::Category::PhysicCommon,
+                     "Error when creating a ConvexMesh: Mesh has a face with less than three vertices.",  __FILE__, __LINE__);
+        }
 
         // Addd the face into the half-edge structure
         mHalfEdgeStructure.addFace(faceVertices);
@@ -123,56 +158,55 @@ bool PolyhedronMesh::createHalfEdgeStructure() {
 
     nbEdges /= 2;
 
-    // If the mesh is valid (check Euler formula V + F - E = 2) and does not use duplicated vertices
-    if (2 + nbEdges - mPolygonVertexArray->getNbFaces() != mPolygonVertexArray->getNbVertices()) {
+    // If the mesh is not valid (check Euler formula V + F - E = 2) (has duplicated vertices)
+    if (2 + nbEdges - polygonVertexArray->getNbFaces() != polygonVertexArray->getNbVertices()) {
 
         RP3D_LOG("PhysicsCommon", Logger::Level::Error, Logger::Category::PhysicCommon,
                  "Error when creating a PolyhedronMesh: input PolygonVertexArray is not valid. Mesh with duplicated vertices is not supported.",  __FILE__, __LINE__);
 
-        assert(false);
+        // Create a new error
+        errors.push_back(Error(std::string("Convex Mesh might have duplicated vertices (invalid Euler formula)")));
 
-        return false;
+        isValid = false;
     }
 
     // Initialize the half-edge structure
     mHalfEdgeStructure.computeHalfEdges();
 
-    // Create the face normals array
-    mFacesNormals = new (mMemoryAllocator.allocate(mHalfEdgeStructure.getNbFaces() * sizeof(Vector3))) Vector3[mHalfEdgeStructure.getNbFaces()];
-
-    return true;
-}
-
-/// Return a vertex
-/**
- * @param index Index of a given vertex in the mesh
- * @return The coordinates of a given vertex in the mesh
- */
-Vector3 PolyhedronMesh::getVertex(uint32 index) const {
-    assert(index < getNbVertices());
-
-    // Get the vertex index in the array with all vertices
-    uint32 vertexIndex = mHalfEdgeStructure.getVertex(index).vertexPointIndex;
-
-    // Return the corresponding vertex from the polygon vertex array
-    return mPolygonVertexArray->getVertex(vertexIndex);
+    return isValid;
 }
 
 // Compute the faces normals
-void PolyhedronMesh::computeFacesNormals() {
+bool PolyhedronMesh::computeFacesNormals(std::vector<Error>& errors) {
+
+    bool isValid = true;
+
+    mFacesNormals.clear();
 
     // For each face
     const uint32 nbFaces = mHalfEdgeStructure.getNbFaces();
     for (uint32 f=0; f < nbFaces; f++) {
+
         const HalfEdgeStructure::Face& face = mHalfEdgeStructure.getFace(f);
 
-        assert(face.faceVertices.size() >= 3);
+        if (face.faceVertices.size() >= 3) {
 
-        const Vector3 vec1 = getVertex(face.faceVertices[1]) - getVertex(face.faceVertices[0]);
-        const Vector3 vec2 = getVertex(face.faceVertices[2]) - getVertex(face.faceVertices[0]);
-        mFacesNormals[f] = vec1.cross(vec2);
-        mFacesNormals[f].normalize();
+            const Vector3 vec1 = getVertex(face.faceVertices[1]) - getVertex(face.faceVertices[0]);
+            const Vector3 vec2 = getVertex(face.faceVertices[2]) - getVertex(face.faceVertices[0]);
+            mFacesNormals.add(vec1.cross(vec2));
+
+            if (mFacesNormals[f].lengthSquare() > MACHINE_EPSILON) {
+
+                mFacesNormals[f].normalize();
+            }
+            else {
+               isValid = false;
+               errors.push_back(Error("Face with index " + std::to_string(f) + " has a zero area"));
+            }
+        }
     }
+
+    return isValid;
 }
 
 // Compute the centroid of the polyhedron
@@ -195,14 +229,14 @@ decimal PolyhedronMesh::getFaceArea(uint32 faceIndex) const {
     const HalfEdgeStructure::Face& face = mHalfEdgeStructure.getFace(faceIndex);
     assert(face.faceVertices.size() >= 3);
 
-    Vector3 v1 = getVertex(face.faceVertices[0]);
+    const Vector3& v1 = getVertex(face.faceVertices[0]);
 
     // For each vertex of the face
     const uint32 nbFaceVertices = static_cast<uint32>(face.faceVertices.size());
     for (uint32 i=2; i < nbFaceVertices; i++) {
 
-        const Vector3 v2 = getVertex(face.faceVertices[i-1]);
-        const Vector3 v3 = getVertex(face.faceVertices[i]);
+        const Vector3& v2 = getVertex(face.faceVertices[i-1]);
+        const Vector3& v3 = getVertex(face.faceVertices[i]);
 
         const Vector3 v1v2 = v2 - v1;
         const Vector3 v1v3 = v3 - v1;
@@ -225,7 +259,7 @@ decimal PolyhedronMesh::getVolume() const {
         const HalfEdgeStructure::Face& face = mHalfEdgeStructure.getFace(f);
         const decimal faceArea = getFaceArea(f);
         const Vector3 faceNormal = mFacesNormals[f];
-        const Vector3 faceVertex = getVertex(face.faceVertices[0]);
+        const Vector3& faceVertex = getVertex(face.faceVertices[0]);
 
         sum += faceVertex.dot(faceNormal) * faceArea;
     }